Split proc_pid.rs into multiple files by area: file, network, bsd, task, work_queue and add a set of helper functions.
