+++ Cargo.toml
@@ -1,11 +1,12 @@
 [package]
 name = "libproc"
-version = "0.4.0"
+version = "0.5.0"
 description = "A rust wrapper of libproc to get information about running processes - currently Mac OS X only"
 authors = ["Andrew Mackenzie <andrew@mackenzie-serres.net>"]
 repository = "https://github.com/andrewdavidmackenzie/libproc-rs"
 readme = "README.md"
 license = "MIT"
+edition = "2018"
 
 [dependencies]
 errno = "^0.1.8"
