+++ src/libproc/mod.rs
@@ -1,2 +1,11 @@
 pub mod proc_pid;
 pub mod kmesg_buffer;
+pub mod work_queue_info;
+pub mod thread_info;
+pub mod task_info;
+pub mod bsd_info;
+pub mod pid_rusage;
+pub mod file_info;
+pub mod net_info;
+
+mod helpers;
