+++ src/libproc/proc_pid.rs
@@ -1,8 +1,12 @@
 extern crate libc;
-extern crate errno;
 
-use self::libc::{uint64_t, int64_t, uint32_t, int32_t, uint16_t, uint8_t, c_short, c_ushort, c_void, c_int, c_uchar, uid_t, gid_t, c_char, off_t, in_addr, sockaddr_un, in6_addr, SOCK_MAXADDRLEN, IF_NAMESIZE};
-use self::errno::errno;
+use crate::libproc::work_queue_info::WorkQueueInfo;
+use crate::libproc::thread_info::ThreadInfo;
+use crate::libproc::task_info::{TaskInfo, TaskAllInfo};
+use crate::libproc::bsd_info::BSDInfo;
+use crate::libproc::helpers;
+
+use self::libc::{uint64_t, uint32_t, c_void, c_int};
 use std::ptr;
 use std::mem;
 
@@ -21,9 +25,6 @@ use std::mem;
 const MAXPATHLEN: usize = 1024;
 const PROC_PIDPATHINFO_MAXSIZE: usize = 4 * MAXPATHLEN;
 
-// from http://opensource.apple.com//source/xnu/xnu-1456.1.26/bsd/sys/proc_info.h
-const MAXTHREADNAMESIZE : usize = 64;
-
 // From http://opensource.apple.com//source/xnu/xnu-1456.1.26/bsd/sys/proc_info.h and
 // http://fxr.watson.org/fxr/source/bsd/sys/proc_info.h?v=xnu-2050.18.24
 #[derive(Copy, Clone)]
@@ -36,132 +37,10 @@ pub enum ProcType {
     ProcPPIDOnly    = 6
 }
 
-// from http://opensource.apple.com//source/xnu/xnu-1504.7.4/bsd/sys/param.h
-const MAXCOMLEN	: usize = 16;
-
 // This trait is needed for polymorphism on pidinfo types, also abstracting flavor in order to provide
 // type-guaranteed flavor correctness
 pub trait PIDInfo: Default {
     fn flavor() -> PidInfoFlavor;
-}
-
-// structures from http://opensource.apple.com//source/xnu/xnu-1456.1.26/bsd/sys/proc_info.h
-#[repr(C)]
-#[derive(Default)]
-pub struct TaskInfo {
-    pub pti_virtual_size        : uint64_t,     // virtual memory size (bytes)
-    pub pti_resident_size       : uint64_t,     // resident memory size (bytes)
-    pub pti_total_user          : uint64_t,     // total time
-    pub pti_total_system        : uint64_t,
-    pub pti_threads_user        : uint64_t,     // existing threads only
-    pub pti_threads_system      : uint64_t,
-    pub pti_policy              : int32_t,      // default policy for new threads
-    pub pti_faults              : int32_t,      // number of page faults
-    pub pti_pageins             : int32_t,      // number of actual pageins
-    pub pti_cow_faults          : int32_t,      // number of copy-on-write faults
-    pub pti_messages_sent       : int32_t,      // number of messages sent
-    pub pti_messages_received   : int32_t,      // number of messages received
-    pub pti_syscalls_mach       : int32_t,      // number of mach system calls
-    pub pti_syscalls_unix       : int32_t,      // number of unix system calls
-    pub pti_csw                 : int32_t,      // number of context switches
-    pub pti_threadnum           : int32_t,      // number of threads in the task
-    pub pti_numrunning          : int32_t,      // number of running threads
-    pub pti_priority            : int32_t       // task priority
-}
-
-impl PIDInfo for TaskInfo {
-    fn flavor() -> PidInfoFlavor { PidInfoFlavor::TaskInfo }
-}
-
-#[repr(C)]
-#[derive(Default)]
-pub struct BSDInfo {
-    pub pbi_flags               : uint32_t,                 // 64bit; emulated etc
-    pub pbi_status              : uint32_t,
-    pub pbi_xstatus             : uint32_t,
-    pub pbi_pid                 : uint32_t,
-    pub pbi_ppid                : uint32_t,
-    pub pbi_uid                 : uid_t,
-    pub pbi_gid                 : gid_t,
-    pub pbi_ruid                : uid_t,
-    pub pbi_rgid                : gid_t,
-    pub pbi_svuid               : uid_t,
-    pub pbi_svgid               : gid_t,
-    pub rfu_1                   : uint32_t,                 // reserved
-    pub pbi_comm                : [c_char; MAXCOMLEN],
-    pub pbi_name                : [c_char; 2 * MAXCOMLEN],  // empty if no name is registered
-    pub pbi_nfiles              : uint32_t,
-    pub pbi_pgid                : uint32_t,
-    pub pbi_pjobc               : uint32_t,
-    pub e_tdev                  : uint32_t,                 // controlling tty dev
-    pub e_tpgid                 : uint32_t,                 // tty process group id
-    pub pbi_nice                : int32_t,
-    pub pbi_start_tvsec         : uint64_t,
-    pub pbi_start_tvusec        : uint64_t
-}
-
-impl PIDInfo for BSDInfo {
-    fn flavor() -> PidInfoFlavor { PidInfoFlavor::TBSDInfo }
-}
-
-#[repr(C)]
-#[derive(Default)]
-pub struct TaskAllInfo {
-    pub pbsd : BSDInfo,
-    pub ptinfo : TaskInfo
-}
-
-impl PIDInfo for TaskAllInfo {
-    fn flavor() -> PidInfoFlavor { PidInfoFlavor::TaskAllInfo }
-}
-
-#[repr(C)]
-pub struct ThreadInfo {
-    pub pth_user_time           : uint64_t,                     // user run time
-    pub pth_system_time         : uint64_t,                     // system run time
-    pub pth_cpu_usage           : int32_t,                      // scaled cpu usage percentage
-    pub pth_policy              : int32_t,                      // scheduling policy in effect
-    pub pth_run_state           : int32_t,                      // run state (see below)
-    pub pth_flags               : int32_t,                      // various flags (see below)
-    pub pth_sleep_time          : int32_t,                      // number of seconds that thread
-    pub pth_curpri              : int32_t,                      // cur priority
-    pub pth_priority            : int32_t,                      // priority
-    pub pth_maxpriority         : int32_t,                      // max priority
-    pub pth_name                : [c_char; MAXTHREADNAMESIZE]   // thread name, if any
-}
-
-impl PIDInfo for ThreadInfo {
-    fn flavor() -> PidInfoFlavor { PidInfoFlavor::ThreadInfo }
-}
-
-impl Default for ThreadInfo {
-    fn default() -> ThreadInfo {
-        ThreadInfo {
-            pth_user_time  : 0,
-            pth_system_time: 0,
-            pth_cpu_usage  : 0,
-            pth_policy     : 0,
-            pth_run_state  : 0,
-            pth_flags      : 0,
-            pth_sleep_time : 0,
-            pth_curpri     : 0,
-            pth_priority   : 0,
-            pth_maxpriority: 0,
-            pth_name       : [0; MAXTHREADNAMESIZE],
-        }
-    }
-}
-
-#[derive(Default)]
-pub struct WorkQueueInfo {
-    pub pwq_nthreads            : uint32_t,     // total number of workqueue threads
-    pub pwq_runthreads          : uint32_t,     // total number of running workqueue threads
-    pub pwq_blockedthreads      : uint32_t,     // total number of blocked workqueue threads
-    pub reserved                : [uint32_t;1]  // reserved for future use
-}
-
-impl PIDInfo for WorkQueueInfo {
-    fn flavor() -> PidInfoFlavor { PidInfoFlavor::WorkQueueInfo }
 }
 
 // From http://opensource.apple.com/source/xnu/xnu-1504.7.4/bsd/kern/proc_info.c
@@ -195,24 +74,18 @@ pub enum PidInfo {
     WorkQueueInfo(WorkQueueInfo)
 }
 
-pub enum PidFDInfoFlavor {
-    VNodeInfo       = 1,
-    VNodePathInfo   = 2,
-    SocketInfo      = 3,
-    PSEMInfo        = 4,
-    PSHMInfo        = 5,
-    PipeInfo        = 6,
-    KQueueInfo      = 7,
-    ATalkInfo       = 8
+// This trait is needed for polymorphism on listpidinfo types, also abstracting flavor in order to provide
+// type-guaranteed flavor correctness
+pub trait ListPIDInfo {
+    type Item;
+    fn flavor() -> PidInfoFlavor;
 }
 
-// From https://opensource.apple.com/source/xnu/xnu-4903.221.2/bsd/sys/resource.h
-pub enum PidRUsageFlavor {
-    V0 = 0,
-    V1 = 1,
-    V2 = 2,
-    V3 = 3,
-    V4 = 4,
+pub struct ListThreads;
+
+impl ListPIDInfo for ListThreads {
+    type Item = uint64_t;
+    fn flavor() -> PidInfoFlavor { PidInfoFlavor::ListThreads }
 }
 
 // this extern block links to the libproc library
@@ -223,8 +96,6 @@ extern {
 
     fn proc_pidinfo(pid : c_int, flavor : c_int, arg: uint64_t, buffer : *mut c_void, buffersize : c_int) -> c_int;
 
-    fn proc_pidfdinfo(pid : c_int, fd : c_int, flavor : c_int, buffer : *mut c_void, buffersize : c_int) -> c_int;
-
     fn proc_name(pid: c_int, buffer: *mut c_void, buffersize: uint32_t) -> c_int;
 
     fn proc_regionfilename(pid: c_int, address: uint64_t, buffer: *mut c_void, buffersize: uint32_t) -> c_int;
@@ -232,16 +103,8 @@ extern {
     fn proc_pidpath(pid: c_int, buffer: *mut c_void, buffersize: uint32_t) -> c_int;
 
     fn proc_libversion(major: *mut c_int, minor: *mut c_int) -> c_int;
-
-    fn proc_pid_rusage(pid: c_int, flavor: c_int, buffer: *mut c_void) -> c_int;
 }
 
-
-pub fn get_errno_with_message(ret: i32) -> String {
-    let e = errno();
-    let code = e.0 as i32;
-    format!("return code = {}, errno = {}, message = '{}'", ret, code, e)
-}
 
 /// Returns the PIDs of the processes active that match the ProcType passed in
 ///
@@ -262,7 +125,7 @@ pub fn get_errno_with_message(ret: i32) -> String {
 pub fn listpids(proc_types: ProcType) -> Result<Vec<u32>, String> {
     let buffer_size = unsafe { proc_listpids(proc_types as u32, 0, ptr::null_mut(), 0) };
     if buffer_size <= 0 {
-        return Err(get_errno_with_message(buffer_size))
+        return Err(helpers::get_errno_with_message(buffer_size))
     }
 
     let capacity = buffer_size as usize / mem::size_of::<u32>();
@@ -272,7 +135,7 @@ pub fn listpids(proc_types: ProcType) -> Result<Vec<u32>, String> {
     let ret = unsafe { proc_listpids(proc_types as u32, 0, buffer_ptr, buffer_size as u32) };
 
     if ret <= 0 {
-        Err(get_errno_with_message(ret))
+        Err(helpers::get_errno_with_message(ret))
     } else {
         let items_count = ret as usize / mem::size_of::<u32>() - 1;
         unsafe {
@@ -282,18 +145,6 @@ pub fn listpids(proc_types: ProcType) -> Result<Vec<u32>, String> {
         Ok(pids)
     }
 }
-
-#[test]
-fn listpids_test() {
-    match listpids(ProcType::ProcAllPIDS) {
-        Ok(pids) => {
-            assert!(pids.len() > 1);
-            println!("Found {} processes using listpids()", pids.len());
-        },
-        Err(err) => assert!(false, "Error listing pids: {}", err)
-    }
-}
-
 
 /// Returns the PIDs of the process that match pid passed in.
 ///
@@ -305,17 +156,16 @@ fn listpids_test() {
 ///
 /// ```
 /// use std::io::Write;
-/// use libproc::libproc::proc_pid::{pidinfo, BSDInfo};
+/// use libproc::libproc::proc_pid::{pidinfo};
+/// use libproc::libproc::bsd_info::{BSDInfo};
+/// use std::process;
 ///
-/// fn pidinfo_test() {
-///     use std::process;
-///     let pid = process::id() as i32;
+/// let pid = process::id() as i32;
 ///
-///     match pidinfo::<BSDInfo>(pid, 0) {
-///         Ok(info) => assert_eq!(info.pbi_pid as i32, pid),
-///         Err(err) => assert!(false, "Error retrieving process info: {}", err)
-///     };
-/// }
+/// match pidinfo::<BSDInfo>(pid, 0) {
+///     Ok(info) => assert_eq!(info.pbi_pid as i32, pid),
+///     Err(err) => assert!(false, "Error retrieving process info: {}", err)
+/// };
 /// ```
 ///
 pub fn pidinfo<T: PIDInfo>(pid : i32, arg: uint64_t) -> Result<T, String> {
@@ -330,100 +180,63 @@ pub fn pidinfo<T: PIDInfo>(pid : i32, arg: uint64_t) -> Result<T, String> {
     };
 
     if ret <= 0 {
-        Err(get_errno_with_message(ret))
+        Err(helpers::get_errno_with_message(ret))
     } else {
         Ok(pidinfo)
     }
 }
 
-#[test]
-fn pidinfo_test() {
-    use std::process;
-    let pid = process::id() as i32;
-
-    match pidinfo::<BSDInfo>(pid, 0) {
-        Ok(info) => assert_eq!(info.pbi_pid as i32, pid),
-        Err(err) => assert!(false, "Error retrieving process info: {}", err)
-    };
-}
-
+///
+/// TODO explain this call or add a link to apple docs that explain it?
+///
+/// # Examples
+/// ```
+/// use libproc::libproc::proc_pid::regionfilename;
+///
+/// // This checks that it can find the regionfilename of the region at address 0, of the init process with PID 1
+/// match regionfilename(1, 0) {
+///     Ok(regionfilename) => println!("Region Filename (at address = 0) of init process PID = 1 is '{}'", regionfilename),
+///     Err(message) => assert!(false, message)
+/// }
+/// ```
+///
 pub fn regionfilename(pid: i32, address: u64) -> Result<String, String> {
-    let mut regionfilenamebuf: Vec<u8> = Vec::with_capacity(PROC_PIDPATHINFO_MAXSIZE - 1);
-    let buffer_ptr = regionfilenamebuf.as_mut_ptr() as *mut c_void;
-    let buffer_size = regionfilenamebuf.capacity() as u32;
+    let mut buf: Vec<u8> = Vec::with_capacity(PROC_PIDPATHINFO_MAXSIZE - 1);
+    let buffer_ptr = buf.as_mut_ptr() as *mut c_void;
+    let buffer_size = buf.capacity() as u32;
     let ret: i32;
 
     unsafe {
         ret = proc_regionfilename(pid, address, buffer_ptr, buffer_size);
     };
 
-    if ret <= 0 {
-        Err(get_errno_with_message(ret))
-    } else {
-        unsafe {
-            regionfilenamebuf.set_len(ret as usize);
-        }
-
-        match String::from_utf8(regionfilenamebuf) {
-            Ok(regionfilename) => Ok(regionfilename),
-            Err(e) => Err(format!("Invalid UTF-8 sequence: {}", e))
-        }
-    }
+    helpers::check_errno(ret, &mut buf)
 }
 
-#[test]
-// This checks that it can find the regionfilename of the region at address 0, of the init process with PID 1
-fn regionfilename_test() {
-    match regionfilename(1, 0) {
-        // run tests with 'cargo test -- --nocapture' to see the test output
-        Ok(regionfilename) => println!("Region Filename (at address = 0) of init process PID = 1 is '{}'", regionfilename),
-        Err(message) => assert!(true, message)
-    }
-}
-
+/// TODO explain this function or link to apple docs that explain it
+///
+/// # Examples
+///
+/// ```
+/// use libproc::libproc::proc_pid::pidpath;
+///
+/// match pidpath(1) {
+///     Ok(path) => println!("Path of init process with PID = 1 is '{}'", path),
+///     Err(message) => assert!(false, message)
+/// }
+/// ```
+///
 pub fn pidpath(pid: i32) -> Result<String, String> {
-    let mut pathbuf: Vec<u8> = Vec::with_capacity(PROC_PIDPATHINFO_MAXSIZE - 1);
-    let buffer_ptr = pathbuf.as_mut_ptr() as *mut c_void;
-    let buffer_size = pathbuf.capacity() as u32;
+    let mut buf: Vec<u8> = Vec::with_capacity(PROC_PIDPATHINFO_MAXSIZE - 1);
+    let buffer_ptr = buf.as_mut_ptr() as *mut c_void;
+    let buffer_size = buf.capacity() as u32;
     let ret: i32;
 
     unsafe {
         ret = proc_pidpath(pid, buffer_ptr, buffer_size);
     };
 
-    if ret <= 0 {
-        Err(get_errno_with_message(ret))
-    } else {
-        unsafe {
-            pathbuf.set_len(ret as usize);
-        }
-
-        match String::from_utf8(pathbuf) {
-            Ok(path) => Ok(path),
-            Err(e) => Err(format!("Invalid UTF-8 sequence: {}", e))
-        }
-    }
-}
-
-#[test]
-// This checks that it can find the path of the init process with PID 1
-fn pidpath_test_init_pid() {
-    match pidpath(1) {
-        // run tests with 'cargo test -- --nocapture' to see the test output
-        Ok(path) => println!("Path of init process with PID = 1 is '{}'", path),
-        Err(message) => assert!(false, message)
-    }
-}
-
-#[test]
-#[should_panic]
-// This checks that it cannot find the path of the process with pid -1
-fn pidpath_test_unknown_pid() {
-    match pidpath(-1) {
-        // run tests with 'cargo test -- --nocapture' to see the test output
-        Ok(path) => assert!(false, "It found the path of process Pwith ID = -1 (path = {}), that's not possible\n", path),
-        Err(message) => assert!(false, message)
-    }
+    helpers::check_errno(ret, &mut buf)
 }
 
 /// Returns the major and minor version numbers of the native librproc library being used
@@ -452,18 +265,7 @@ pub fn libversion() -> Result<(i32, i32), String> {
     if ret == 0 {
         Ok((major, minor))
     } else {
-        Err(get_errno_with_message(ret))
-    }
-}
-
-#[test]
-fn libversion_test() {
-    match libversion() {
-        Ok((major, minor)) => {
-            // run tests with 'cargo test -- --nocapture' to see the test output
-            println!("Major = {}, Minor = {}", major, minor);
-        },
-        Err(message) => assert!(false, message)
+        Err(helpers::get_errno_with_message(ret))
     }
 }
 
@@ -491,7 +293,7 @@ pub fn name(pid: i32) -> Result<String, String> {
     };
 
     if ret <= 0 {
-        Err(get_errno_with_message(ret))
+        Err(helpers::get_errno_with_message(ret))
     } else {
         unsafe {
             namebuf.set_len(ret as usize);
@@ -504,24 +306,6 @@ pub fn name(pid: i32) -> Result<String, String> {
     }
 }
 
-#[test]
-// error: Process didn't exit successfully: `/Users/andrew/workspace/libproc-rs/target/debug/libproc-503ad0ba07eb6318` (signal: 11, SIGSEGV: invalid memory reference)
-// This checks that it can find the name of the init process with PID 1
-fn name_test_init_pid() {
-    match pidpath(1) {
-        // run tests with 'cargo test -- --nocapture' to see the test output
-        Ok(path) => println!("Name of init process PID = 1 is '{}'", path),
-        Err(message) => assert!(true, message)
-    }
-}
-
-// This trait is needed for polymorphism on listpidinfo types, also abstracting flavor in order to provide
-// type-guaranteed flavor correctness
-pub trait ListPIDInfo {
-    type Item;
-    fn flavor() -> PidInfoFlavor;
-}
-
 /// Returns the information of the process that match pid passed in.
 /// `max_len` is the maximum number of array to return.
 /// The length of return value: `Vec<T::Item>` may be less than `max_len`.
@@ -530,7 +314,9 @@ pub trait ListPIDInfo {
 ///
 /// ```
 /// use std::io::Write;
-/// use libproc::libproc::proc_pid::{listpidinfo, pidinfo, ListFDs, TaskAllInfo, ProcFDType};
+/// use libproc::libproc::proc_pid::{listpidinfo, pidinfo};
+/// use libproc::libproc::task_info::{TaskAllInfo};
+/// use libproc::libproc::file_info::{ListFDs, ProcFDType};
 ///
 /// fn listpidinfo_test() {
 ///     use std::process;
@@ -562,7 +348,7 @@ pub fn listpidinfo<T: ListPIDInfo>(pid : i32, max_len: usize) -> Result<Vec<T::I
     };
 
     if ret <= 0 {
-        Err(get_errno_with_message(ret))
+        Err(helpers::get_errno_with_message(ret))
     } else {
         let actual_len = ret as usize / mem::size_of::<T::Item>();
         buffer.truncate(actual_len);
@@ -570,725 +356,75 @@ pub fn listpidinfo<T: ListPIDInfo>(pid : i32, max_len: usize) -> Result<Vec<T::I
     }
 }
 
-#[test]
-fn listpidinfo_test() {
-    use std::process;
-    let pid = process::id() as i32;
+#[cfg(test)]
+mod test {
+    use super::{pidinfo, listpidinfo, ListThreads, pidpath, libversion};
+    use crate::libproc::bsd_info::BSDInfo;
+    use crate::libproc::task_info::TaskAllInfo;
+    use crate::libproc::file_info::ListFDs;
 
-    match pidinfo::<TaskAllInfo>(pid, 0) {
-        Ok(info) => {
-            match listpidinfo::<ListThreads>(pid, info.ptinfo.pti_threadnum as usize) {
-                Ok(threads) => assert!(threads.len()>0),
-                Err(err) => assert!(false, "Error retrieving process info: {}", err)
-            }
-            match listpidinfo::<ListFDs>(pid, info.pbsd.pbi_nfiles as usize) {
-                Ok(fds) => assert!(fds.len()>0),
-                Err(err) => assert!(false, "Error retrieving process info: {}", err)
-            }
-        },
-        Err(err) => assert!(false, "Error retrieving process info: {}", err)
-    };
-}
+    #[test]
+    fn pidinfo_test() {
+        use std::process;
+        let pid = process::id() as i32;
 
-pub struct ListThreads;
-
-impl ListPIDInfo for ListThreads {
-    type Item = uint64_t;
-    fn flavor() -> PidInfoFlavor { PidInfoFlavor::ListThreads }
-}
-
-pub struct ListFDs;
-
-impl ListPIDInfo for ListFDs {
-    type Item = ProcFDInfo;
-    fn flavor() -> PidInfoFlavor { PidInfoFlavor::ListFDs }
-}
-
-#[repr(C)]
-pub struct ProcFDInfo {
-    pub proc_fd: int32_t,
-    pub proc_fdtype: uint32_t,
-}
-
-#[derive(Copy, Clone, Debug)]
-pub enum ProcFDType {
-    /// AppleTalk
-    ATalk    = 0,
-    /// Vnode
-    VNode    = 1,
-    /// Socket
-    Socket   = 2,
-    /// POSIX shared memory
-    PSHM     = 3,
-    /// POSIX semaphore
-    PSEM     = 4,
-    /// Kqueue
-    KQueue   = 5,
-    /// Pipe
-    Pipe     = 6,
-    /// FSEvents
-    FSEvents = 7,
-    /// Unknown
-    Unknown,
-}
-
-impl From<uint32_t> for ProcFDType {
-    fn from(value: uint32_t) -> ProcFDType {
-        match value {
-            0 => ProcFDType::ATalk   ,
-            1 => ProcFDType::VNode   ,
-            2 => ProcFDType::Socket  ,
-            3 => ProcFDType::PSHM    ,
-            4 => ProcFDType::PSEM    ,
-            5 => ProcFDType::KQueue  ,
-            6 => ProcFDType::Pipe    ,
-            7 => ProcFDType::FSEvents,
-            _ => ProcFDType::Unknown ,
-        }
+        match pidinfo::<BSDInfo>(pid, 0) {
+            Ok(info) => assert_eq!(info.pbi_pid as i32, pid),
+            Err(err) => assert!(false, "Error retrieving process info: {}", err)
+        };
     }
-}
 
-// This trait is needed for polymorphism on pidfdinfo types, also abstracting flavor in order to provide
-// type-guaranteed flavor correctness
-pub trait PIDFDInfo: Default {
-    fn flavor() -> PidFDInfoFlavor;
-}
+    #[test]
+    fn listpidinfo_test() {
+        use std::process;
+        let pid = process::id() as i32;
 
-/// Returns the information about file descriptors of the process that match pid passed in.
-///
-/// # Examples
-///
-/// ```
-/// use std::io::Write;
-/// use std::net::TcpListener;
-/// use libproc::libproc::proc_pid::{listpidinfo, pidinfo, pidfdinfo, ListFDs, ListThreads, BSDInfo, ProcFDType, SocketFDInfo, SocketInfoKind};
-///
-/// fn pidfdinfo_test() {
-///     use std::process;
-///     let pid = process::id() as i32;
-///
-///     // Open TCP port:8000 to test.
-///     let _listener = TcpListener::bind("127.0.0.1:8000");
-///
-///     if let Ok(info) = pidinfo::<BSDInfo>(pid, 0) {
-///         if let Ok(fds) = listpidinfo::<ListFDs>(pid, info.pbi_nfiles as usize) {
-///             for fd in &fds {
-///                 match fd.proc_fdtype.into() {
-///                     ProcFDType::Socket => {
-///                         if let Ok(socket) = pidfdinfo::<SocketFDInfo>(pid, fd.proc_fd) {
-///                             match socket.psi.soi_kind.into() {
-///                                 SocketInfoKind::Tcp => {
-///                                     // access to the member of `soi_proto` is unsafe becasuse of union type.
-///                                     let info = unsafe { socket.psi.soi_proto.pri_tcp };
-///
-///                                     // change endian and cut off because insi_lport is network endian and 16bit witdh.
-///                                     let mut port = 0;
-///                                     port |= info.tcpsi_ini.insi_lport >> 8 & 0x00ff;
-///                                     port |= info.tcpsi_ini.insi_lport << 8 & 0xff00;
-///
-///                                     // access to the member of `insi_laddr` is unsafe becasuse of union type.
-///                                     let s_addr = unsafe { info.tcpsi_ini.insi_laddr.ina_46.i46a_addr4.s_addr };
-///
-///                                     // change endian because insi_laddr is network endian.
-///                                     let mut addr = 0;
-///                                     addr |= s_addr >> 24 & 0x000000ff;
-///                                     addr |= s_addr >> 8  & 0x0000ff00;
-///                                     addr |= s_addr << 8  & 0x00ff0000;
-///                                     addr |= s_addr << 24 & 0xff000000;
-///
-///                                     println!("{}.{}.{}.{}:{}", addr >> 24 & 0xff, addr >> 16 & 0xff, addr >> 8 & 0xff, addr & 0xff, port);
-///                                 },
-///                                 _ => (),
-///                             }
-///                         }
-///                     },
-///                     _ => (),
-///                 }
-///             }
-///         }
-///     }
-/// }
-/// ```
-///
-pub fn pidfdinfo<T: PIDFDInfo>(pid : i32, fd: int32_t) -> Result<T, String> {
-    let flavor = T::flavor() as i32;
-    let buffer_size = mem::size_of::<T>() as i32;
-    let mut pidinfo = T::default();
-    let buffer_ptr = &mut pidinfo as *mut _ as *mut c_void;
-    let ret: i32;
-
-    unsafe {
-        ret = proc_pidfdinfo(pid, fd, flavor, buffer_ptr, buffer_size);
-    };
-
-    if ret <= 0 {
-        Err(get_errno_with_message(ret))
-    } else {
-        Ok(pidinfo)
-    }
-}
-
-#[test]
-fn pidfdinfo_test() {
-    use std::process;
-    use std::net::TcpListener;
-    let pid = process::id() as i32;
-
-    let _listener = TcpListener::bind("127.0.0.1:65535");
-
-    let info = pidinfo::<BSDInfo>(pid, 0).unwrap();
-    let fds = listpidinfo::<ListFDs>(pid, info.pbi_nfiles as usize).unwrap();
-    for fd in fds {
-        match fd.proc_fdtype.into() {
-            ProcFDType::Socket => {
-                let socket = pidfdinfo::<SocketFDInfo>(pid, fd.proc_fd).unwrap();
-                match socket.psi.soi_kind.into() {
-                    SocketInfoKind::Tcp => unsafe {
-                        let info = socket.psi.soi_proto.pri_tcp;
-                        assert_eq!(socket.psi.soi_protocol, libc::IPPROTO_TCP);
-                        assert_eq!(info.tcpsi_ini.insi_lport as u32, 65535);
-                    }
-                    _ => (),
+        match pidinfo::<TaskAllInfo>(pid, 0) {
+            Ok(info) => {
+                match listpidinfo::<ListThreads>(pid, info.ptinfo.pti_threadnum as usize) {
+                    Ok(threads) => assert!(threads.len()>0),
+                    Err(err) => assert!(false, "Error retrieving process info: {}", err)
+                }
+                match listpidinfo::<ListFDs>(pid, info.pbsd.pbi_nfiles as usize) {
+                    Ok(fds) => assert!(fds.len()>0),
+                    Err(err) => assert!(false, "Error retrieving process info: {}", err)
                 }
             },
-            _ => (),
+            Err(err) => assert!(false, "Error retrieving process info: {}", err)
+        };
+    }
+
+    #[test]
+    // error: Process didn't exit successfully:
+    // `/Users/andrew/workspace/libproc-rs/target/debug/libproc-503ad0ba07eb6318` (signal: 11, SIGSEGV: invalid memory reference)
+    // This checks that it can find the name of the init process with PID 1
+    fn name_test_init_pid() {
+        match pidpath(1) {
+            // run tests with 'cargo test -- --nocapture' to see the test output
+            Ok(path) => println!("Name of init process PID = 1 is '{}'", path),
+            Err(message) => assert!(true, message)
         }
     }
-}
 
-#[repr(C)]
-#[derive(Default)]
-pub struct SocketFDInfo {
-    pub pfi: ProcFileInfo,
-    pub psi: SocketInfo,
-}
-
-impl PIDFDInfo for SocketFDInfo {
-    fn flavor() -> PidFDInfoFlavor { PidFDInfoFlavor::SocketInfo }
-}
-
-#[repr(C)]
-#[derive(Default)]
-pub struct ProcFileInfo {
-    pub fi_openflags: uint32_t,
-    pub fi_status   : uint32_t,
-    pub fi_offset   : off_t,
-    pub fi_type     : int32_t,
-    pub rfu_1       : int32_t,
-}
-
-#[derive(Copy, Clone, Debug)]
-pub enum SocketInfoKind {
-    Generic   = 0,
-    /// IPv4 and IPv6 Sockets
-    In        = 1,
-    /// TCP Sockets
-    Tcp       = 2,
-    /// Unix Domain Sockets
-    Un        = 3,
-    /// PF_NDRV Sockets
-    Ndrv      = 4,
-    /// Kernel Event Sockets
-    KernEvent = 5,
-    /// Kernel Control Sockets
-    KernCtl   = 6,
-    /// Unknown
-    Unknown,
-}
-
-impl From<c_int> for SocketInfoKind {
-    fn from(value: c_int) -> SocketInfoKind {
-        match value {
-            0 => SocketInfoKind::Generic  ,
-            1 => SocketInfoKind::In       ,
-            2 => SocketInfoKind::Tcp      ,
-            3 => SocketInfoKind::Un       ,
-            4 => SocketInfoKind::Ndrv     ,
-            5 => SocketInfoKind::KernEvent,
-            6 => SocketInfoKind::KernCtl  ,
-            _ => SocketInfoKind::Unknown  ,
+    #[test]
+    fn libversion_test() {
+        match libversion() {
+            Ok((major, minor)) => {
+                // run tests with 'cargo test -- --nocapture' to see the test output
+                println!("Major = {}, Minor = {}", major, minor);
+            },
+            Err(message) => assert!(false, message)
         }
     }
-}
 
-#[repr(C)]
-#[derive(Default)]
-pub struct SocketInfo {
-    pub soi_stat    : VInfoStat,
-    pub soi_so      : uint64_t,
-    pub soi_pcb     : uint64_t,
-    pub soi_type    : c_int,
-    pub soi_protocol: c_int,
-    pub soi_family  : c_int,
-    pub soi_options : c_short,
-    pub soi_linger  : c_short,
-    pub soi_state   : c_short,
-    pub soi_qlen    : c_short,
-    pub soi_incqlen : c_short,
-    pub soi_qlimit  : c_short,
-    pub soi_timeo   : c_short,
-    pub soi_error   : c_ushort,
-    pub soi_oobmark : uint32_t,
-    pub soi_rcv     : SockBufInfo,
-    pub soi_snd     : SockBufInfo,
-    pub soi_kind    : c_int,
-    pub rfu_1       : uint32_t,
-    pub soi_proto   : SocketInfoProto,
-}
-
-#[repr(C)]
-#[derive(Default)]
-pub struct VInfoStat {
-    pub vst_dev          : uint32_t,
-    pub vst_mode         : uint16_t,
-    pub vst_nlink        : uint16_t,
-    pub vst_ino          : uint64_t,
-    pub vst_uid          : uid_t,
-    pub vst_gid          : gid_t,
-    pub vst_atime        : int64_t,
-    pub vst_atimensec    : int64_t,
-    pub vst_mtime        : int64_t,
-    pub vst_mtimensec    : int64_t,
-    pub vst_ctime        : int64_t,
-    pub vst_ctimensec    : int64_t,
-    pub vst_birthtime    : int64_t,
-    pub vst_birthtimensec: int64_t,
-    pub vst_size         : off_t,
-    pub vst_blocks       : int64_t,
-    pub vst_blksize      : int32_t,
-    pub vst_flags        : uint32_t,
-    pub vst_gen          : uint32_t,
-    pub vst_rdev         : uint32_t,
-    pub vst_qspare       : [int64_t; 2],
-}
-
-#[repr(C)]
-#[derive(Default)]
-pub struct SockBufInfo {
-    pub sbi_cc   : uint32_t,
-    pub sbi_hiwat: uint32_t,
-    pub sbi_mbcnt: uint32_t,
-    pub sbi_mbmax: uint32_t,
-    pub sbi_lowat: uint32_t,
-    pub sbi_flags: c_short,
-    pub sbi_timeo: c_short,
-}
-
-#[repr(C)]
-pub union SocketInfoProto {
-    pub pri_in        : InSockInfo,
-    pub pri_tcp       : TcpSockInfo,
-    pub pri_un        : UnSockInfo,
-    pub pri_ndrv      : NdrvInfo,
-    pub pri_kern_event: KernEventInfo,
-    pub pri_kern_ctl  : KernCtlInfo,
-}
-
-impl Default for SocketInfoProto {
-    fn default() -> SocketInfoProto {
-        SocketInfoProto {
-            pri_in: Default::default(),
+    #[test]
+    #[should_panic]
+    // This checks that it cannot find the path of the process with pid -1
+    fn pidpath_test_unknown_pid() {
+        match pidpath(-1) {
+            // run tests with 'cargo test -- --nocapture' to see the test output
+            Ok(path) => assert!(false, "It found the path of process Pwith ID = -1 (path = {}), that's not possible\n", path),
+            Err(message) => assert!(false, message)
         }
     }
-}
-
-#[repr(C)]
-#[derive(Copy, Clone)]
-pub struct In4In6Addr {
-    pub i46a_pad32: [uint32_t; 3],
-    pub i46a_addr4: in_addr,
-}
-
-impl Default for In4In6Addr {
-    fn default() -> In4In6Addr {
-        In4In6Addr {
-            i46a_pad32: [0; 3],
-            i46a_addr4: in_addr{s_addr: 0},
-        }
-    }
-}
-
-#[repr(C)]
-#[derive(Copy, Clone, Default)]
-pub struct InSockInfo {
-    pub insi_fport : c_int,
-    pub insi_lport : c_int,
-    pub insi_gencnt: uint64_t,
-    pub insi_flags : uint32_t,
-    pub insi_flow  : uint32_t,
-    pub insi_vflag : uint8_t,
-    pub insi_ip_ttl: uint8_t,
-    pub rfu_1      : uint32_t,
-    pub insi_faddr : InSIAddr,
-    pub insi_laddr : InSIAddr,
-    pub insi_v4    : InSIV4,
-    pub insi_v6    : InSIV6,
-}
-
-#[repr(C)]
-#[derive(Copy, Clone, Default)]
-pub struct InSIV4 {
-    pub in4_top: c_uchar,
-}
-
-#[repr(C)]
-#[derive(Copy, Clone, Default)]
-pub struct InSIV6 {
-    pub in6_hlim   : uint8_t,
-    pub in6_cksum  : c_int,
-    pub in6_ifindex: c_ushort,
-    pub in6_hops   : c_short,
-}
-
-#[repr(C)]
-#[derive(Copy, Clone)]
-pub union InSIAddr {
-    pub ina_46: In4In6Addr,
-    pub ina_6 : in6_addr,
-}
-
-impl Default for InSIAddr {
-    fn default() -> InSIAddr {
-        InSIAddr {
-            ina_46: Default::default(),
-        }
-    }
-}
-
-#[derive(Copy, Clone, Debug)]
-pub enum TcpSIState {
-    /// Closed
-    Closed      = 0,
-    /// Listening for connection
-    Listen      = 1,
-    /// Active, have sent syn
-    SynSent     = 2,
-    /// Have send and received syn
-    SynReceived = 3,
-    /// Established
-    Established = 4,
-    /// Rcvd fin, waiting for close
-    CloseWait   = 5,
-    /// Have closed, sent fin
-    FinWait1    = 6,
-    /// Closed xchd FIN; await FIN ACK
-    Closing     = 7,
-    /// Had fin and close; await FIN ACK
-    LastAck     = 8,
-    /// Have closed, fin is acked
-    FinWait2    = 9,
-    /// In 2*msl quiet wait after close
-    TimeWait    = 10,
-    /// Pseudo state: reserved
-    Reserved    = 11,
-    /// Unknown
-    Unknown,
-}
-
-impl From<c_int> for TcpSIState {
-    fn from(value: c_int) -> TcpSIState {
-        match value {
-            0  => TcpSIState::Closed     ,
-            1  => TcpSIState::Listen     ,
-            2  => TcpSIState::SynSent    ,
-            3  => TcpSIState::SynReceived,
-            4  => TcpSIState::Established,
-            5  => TcpSIState::CloseWait  ,
-            6  => TcpSIState::FinWait1   ,
-            7  => TcpSIState::Closing    ,
-            8  => TcpSIState::LastAck    ,
-            9  => TcpSIState::FinWait2   ,
-            10 => TcpSIState::TimeWait   ,
-            11 => TcpSIState::Reserved   ,
-            _  => TcpSIState::Unknown    ,
-        }
-    }
-}
-
-const TSI_T_NTIMERS : usize = 4;
-
-#[repr(C)]
-#[derive(Copy, Clone, Default)]
-pub struct TcpSockInfo {
-    pub tcpsi_ini  : InSockInfo,
-    pub tcpsi_state: c_int,
-    pub tcpsi_timer: [c_int; TSI_T_NTIMERS],
-    pub tcpsi_mss  : c_int,
-    pub tcpsi_flags: uint32_t,
-    pub rfu_1      : uint32_t,
-    pub tcpsi_tp   : uint64_t,
-}
-
-#[repr(C)]
-#[derive(Copy, Clone, Default)]
-pub struct UnSockInfo {
-    pub unsi_conn_so : uint64_t,
-    pub unsi_conn_pcb: uint64_t,
-    pub unsi_addr    : UnSIAddr,
-    pub unsi_caddr   : UnSIAddr,
-}
-
-#[repr(C)]
-#[derive(Copy, Clone)]
-pub union UnSIAddr {
-    pub ua_sun  : sockaddr_un,
-    pub ua_dummy: [c_char; SOCK_MAXADDRLEN as usize],
-}
-
-impl Default for UnSIAddr {
-    fn default() -> UnSIAddr {
-        UnSIAddr {
-            ua_dummy: [0; SOCK_MAXADDRLEN as usize],
-        }
-    }
-}
-
-#[repr(C)]
-#[derive(Copy, Clone, Default)]
-pub struct NdrvInfo {
-    pub ndrvsi_if_family: uint32_t,
-    pub ndrvsi_if_unit  : uint32_t,
-    pub ndrvsi_if_name  : [c_char; IF_NAMESIZE],
-}
-
-#[repr(C)]
-#[derive(Copy, Clone, Default)]
-pub struct KernEventInfo {
-    pub kesi_vendor_code_filter: uint32_t,
-    pub kesi_class_filter      : uint32_t,
-    pub kesi_subclass_filter   : uint32_t,
-}
-
-const MAX_KCTL_NAME : usize = 96;
-
-#[repr(C)]
-#[derive(Copy, Clone)]
-pub struct KernCtlInfo {
-    pub kcsi_id         : uint32_t,
-    pub kcsi_reg_unit   : uint32_t,
-    pub kcsi_flags      : uint32_t,
-    pub kcsi_recvbufsize: uint32_t,
-    pub kcsi_sendbufsize: uint32_t,
-    pub kcsi_unit       : uint32_t,
-    pub kcsi_name       : [c_char; MAX_KCTL_NAME],
-}
-
-impl Default for KernCtlInfo {
-    fn default() -> KernCtlInfo {
-        KernCtlInfo {
-            kcsi_id         : 0,
-            kcsi_reg_unit   : 0,
-            kcsi_flags      : 0,
-            kcsi_recvbufsize: 0,
-            kcsi_sendbufsize: 0,
-            kcsi_unit       : 0,
-            kcsi_name       : [0; MAX_KCTL_NAME],
-        }
-    }
-}
-
-// This trait is needed for polymorphism on pidrusage types, also abstracting flavor in order to provide
-// type-guaranteed flavor correctness
-pub trait PIDRUsage: Default {
-    fn flavor() -> PidRUsageFlavor;
-}
-
-/// Returns the information about resources of the process that match pid passed in.
-///
-/// # Examples
-///
-/// ```
-/// use std::io::Write;
-/// use libproc::libproc::proc_pid::{pidrusage, RUsageInfoV2};
-///
-/// fn pidrusage_test() {
-///     use std::process;
-///     let pid = process::id() as i32;
-///
-///     if let Ok(res) = pidrusage::<RUsageInfoV2>(pid) {
-///         println!("UUID: {:?}, Disk Read: {}, Disk Write: {}", res.ri_uuid, res.ri_diskio_bytesread, res.ri_diskio_byteswritten);
-///     }
-/// }
-/// ```
-///
-pub fn pidrusage<T: PIDRUsage>(pid : i32) -> Result<T, String> {
-    let flavor = T::flavor() as i32;
-    let mut pidrusage = T::default();
-    let buffer_ptr = &mut pidrusage as *mut _ as *mut c_void;
-    let ret: i32;
-
-    unsafe {
-        ret = proc_pid_rusage(pid, flavor, buffer_ptr);
-    };
-
-    if ret < 0 {
-        Err(get_errno_with_message(ret))
-    } else {
-        Ok(pidrusage)
-    }
-}
-
-#[test]
-fn pidrusage_test() {
-    use std::process;
-    let pid = process::id() as i32;
-
-    let res = pidrusage::<RUsageInfoV2>(pid).unwrap();
-    println!("UUID: {:?}, Disk Read: {}, Disk Write: {}", res.ri_uuid, res.ri_diskio_bytesread, res.ri_diskio_byteswritten);
-}
-
-#[repr(C)]
-#[derive(Default)]
-pub struct RUsageInfoV0 {
-    pub ri_uuid              : [uint8_t; 16],
-    pub ri_user_time         : uint64_t,
-    pub ri_system_time       : uint64_t,
-    pub ri_pkg_idle_wkups    : uint64_t,
-    pub ri_interrupt_wkups   : uint64_t,
-    pub ri_pageins           : uint64_t,
-    pub ri_wired_size        : uint64_t,
-    pub ri_resident_size     : uint64_t,
-    pub ri_phys_footprint    : uint64_t,
-    pub ri_proc_start_abstime: uint64_t,
-    pub ri_proc_exit_abstime : uint64_t,
-}
-
-impl PIDRUsage for RUsageInfoV0 {
-    fn flavor() -> PidRUsageFlavor { PidRUsageFlavor::V0 }
-}
-
-#[repr(C)]
-#[derive(Default)]
-pub struct RUsageInfoV1 {
-    pub ri_uuid                 : [uint8_t; 16],
-    pub ri_user_time            : uint64_t,
-    pub ri_system_time          : uint64_t,
-    pub ri_pkg_idle_wkups       : uint64_t,
-    pub ri_interrupt_wkups      : uint64_t,
-    pub ri_pageins              : uint64_t,
-    pub ri_wired_size           : uint64_t,
-    pub ri_resident_size        : uint64_t,
-    pub ri_phys_footprint       : uint64_t,
-    pub ri_proc_start_abstime   : uint64_t,
-    pub ri_proc_exit_abstime    : uint64_t,
-    pub ri_child_user_time      : uint64_t,
-    pub ri_child_system_time    : uint64_t,
-    pub ri_child_pkg_idle_wkups : uint64_t,
-    pub ri_child_interrupt_wkups: uint64_t,
-    pub ri_child_pageins        : uint64_t,
-    pub ri_child_elapsed_abstime: uint64_t,
-}
-
-impl PIDRUsage for RUsageInfoV1 {
-    fn flavor() -> PidRUsageFlavor { PidRUsageFlavor::V1 }
-}
-
-#[repr(C)]
-#[derive(Debug, Default)]
-pub struct RUsageInfoV2 {
-    pub ri_uuid                 : [uint8_t; 16],
-    pub ri_user_time            : uint64_t,
-    pub ri_system_time          : uint64_t,
-    pub ri_pkg_idle_wkups       : uint64_t,
-    pub ri_interrupt_wkups      : uint64_t,
-    pub ri_pageins              : uint64_t,
-    pub ri_wired_size           : uint64_t,
-    pub ri_resident_size        : uint64_t,
-    pub ri_phys_footprint       : uint64_t,
-    pub ri_proc_start_abstime   : uint64_t,
-    pub ri_proc_exit_abstime    : uint64_t,
-    pub ri_child_user_time      : uint64_t,
-    pub ri_child_system_time    : uint64_t,
-    pub ri_child_pkg_idle_wkups : uint64_t,
-    pub ri_child_interrupt_wkups: uint64_t,
-    pub ri_child_pageins        : uint64_t,
-    pub ri_child_elapsed_abstime: uint64_t,
-    pub ri_diskio_bytesread     : uint64_t,
-    pub ri_diskio_byteswritten  : uint64_t,
-}
-
-impl PIDRUsage for RUsageInfoV2 {
-    fn flavor() -> PidRUsageFlavor { PidRUsageFlavor::V2 }
-}
-
-#[repr(C)]
-#[derive(Default)]
-pub struct RUsageInfoV3 {
-    pub ri_uuid                         : [uint8_t; 16],
-    pub ri_user_time                    : uint64_t,
-    pub ri_system_time                  : uint64_t,
-    pub ri_pkg_idle_wkups               : uint64_t,
-    pub ri_interrupt_wkups              : uint64_t,
-    pub ri_pageins                      : uint64_t,
-    pub ri_wired_size                   : uint64_t,
-    pub ri_resident_size                : uint64_t,
-    pub ri_phys_footprint               : uint64_t,
-    pub ri_proc_start_abstime           : uint64_t,
-    pub ri_proc_exit_abstime            : uint64_t,
-    pub ri_child_user_time              : uint64_t,
-    pub ri_child_system_time            : uint64_t,
-    pub ri_child_pkg_idle_wkups         : uint64_t,
-    pub ri_child_interrupt_wkups        : uint64_t,
-    pub ri_child_pageins                : uint64_t,
-    pub ri_child_elapsed_abstime        : uint64_t,
-    pub ri_diskio_bytesread             : uint64_t,
-    pub ri_diskio_byteswritten          : uint64_t,
-    pub ri_cpu_time_qos_default         : uint64_t,
-    pub ri_cpu_time_qos_maintenance     : uint64_t,
-    pub ri_cpu_time_qos_background      : uint64_t,
-    pub ri_cpu_time_qos_utility         : uint64_t,
-    pub ri_cpu_time_qos_legacy          : uint64_t,
-    pub ri_cpu_time_qos_user_initiated  : uint64_t,
-    pub ri_cpu_time_qos_user_interactive: uint64_t,
-    pub ri_billed_system_time           : uint64_t,
-    pub ri_serviced_system_time         : uint64_t,
-}
-
-impl PIDRUsage for RUsageInfoV3 {
-    fn flavor() -> PidRUsageFlavor { PidRUsageFlavor::V3 }
-}
-
-#[repr(C)]
-#[derive(Default)]
-pub struct RUsageInfoV4 {
-    pub ri_uuid                         : [uint8_t; 16],
-    pub ri_user_time                    : uint64_t,
-    pub ri_system_time                  : uint64_t,
-    pub ri_pkg_idle_wkups               : uint64_t,
-    pub ri_interrupt_wkups              : uint64_t,
-    pub ri_pageins                      : uint64_t,
-    pub ri_wired_size                   : uint64_t,
-    pub ri_resident_size                : uint64_t,
-    pub ri_phys_footprint               : uint64_t,
-    pub ri_proc_start_abstime           : uint64_t,
-    pub ri_proc_exit_abstime            : uint64_t,
-    pub ri_child_user_time              : uint64_t,
-    pub ri_child_system_time            : uint64_t,
-    pub ri_child_pkg_idle_wkups         : uint64_t,
-    pub ri_child_interrupt_wkups        : uint64_t,
-    pub ri_child_pageins                : uint64_t,
-    pub ri_child_elapsed_abstime        : uint64_t,
-    pub ri_diskio_bytesread             : uint64_t,
-    pub ri_diskio_byteswritten          : uint64_t,
-    pub ri_cpu_time_qos_default         : uint64_t,
-    pub ri_cpu_time_qos_maintenance     : uint64_t,
-    pub ri_cpu_time_qos_background      : uint64_t,
-    pub ri_cpu_time_qos_utility         : uint64_t,
-    pub ri_cpu_time_qos_legacy          : uint64_t,
-    pub ri_cpu_time_qos_user_initiated  : uint64_t,
-    pub ri_cpu_time_qos_user_interactive: uint64_t,
-    pub ri_billed_system_time           : uint64_t,
-    pub ri_serviced_system_time         : uint64_t,
-    pub ri_logical_writes               : uint64_t,
-    pub ri_lifetime_max_phys_footprint  : uint64_t,
-    pub ri_instructions                 : uint64_t,
-    pub ri_cycles                       : uint64_t,
-    pub ri_billed_energy                : uint64_t,
-    pub ri_serviced_energy              : uint64_t,
-    pub ri_interval_max_phys_footprint  : uint64_t,
-    pub ri_unused                       : [uint64_t; 1],
-}
-
-impl PIDRUsage for RUsageInfoV4 {
-    fn flavor() -> PidRUsageFlavor { PidRUsageFlavor::V4 }
 }
